requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
distro==1.6.0
Django==3.2.8
et-xmlfile==1.1.0
idna==3.3
joblib==1.1.0
langdetect==1.0.9
libretranslatepy==2.1.1
lxml==4.6.4
nltk==3.6.5
numpy==1.21.3
openpyxl==3.0.9
pandas==1.3.4
Pillow==8.4.0
PyMuPDF==1.19.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
regex==2021.10.23
requests==2.26.0
six==1.16.0
soupsieve==2.3
sqlparse==0.4.2
tabula-py==2.3.0
textblob==0.17.1
tqdm==4.62.3
translate==3.6.1
urllib3==1.26.7
xlrd==2.0.1
